Refactor all other tabs for bot form
