+++ web/src/components/TakeProfitTab.jsx
